Declared styled-components as a peer dependency (#29)
It was bundling styled-components inside vis-core.

That creates a second ThemeContext, so the app’s can’t reach vis-core and theme.mq.mobile is undefined.

So moved the styled components to the peerDep.

Another fix: replaced p tag to div

as Vite throwing error that p cannot be a descendant of p tag